+++ src/arctic_route_risk/__init__.py
@@ -30,11 +30,27 @@ from arctic_route_risk.modeling import (
     RiskModelBackend,
     RiskModelInput,
     RiskModelOutput,
+    RiskTrainingEvaluation,
+    RiskTrainingReadiness,
+    RiskTrainingSample,
+    RiskTrainingSplit,
     RuleBaselineBackend,
+    build_one_hour_training_samples,
+    evaluate_persistence_baseline,
     intake_legacy_cnn_zip,
+    iter_training_sample_documents,
+    readiness_as_dict,
+    summarize_training_readiness,
+    temporal_holdout_split,
 )
 from arctic_route_risk.publishing import PersistentRiskStore
 from arctic_route_risk.service import RiskBuildRequest, RiskBuildService
+from arctic_route_risk.single_factor import (
+    SingleFactorOutputPaths,
+    SingleFactorRiskLayer,
+    build_realtime_single_factor_layers,
+    write_realtime_single_factor_outputs,
+)
 
 __all__ = [
     "MODEL_ARTIFACT_INVALID",
@@ -61,13 +77,27 @@ __all__ = [
     "RiskModelInput",
     "RiskModelOutput",
     "RiskPipelineError",
+    "RiskTrainingEvaluation",
+    "RiskTrainingReadiness",
+    "RiskTrainingSample",
+    "RiskTrainingSplit",
     "RuleBaselineBackend",
+    "SingleFactorOutputPaths",
+    "SingleFactorRiskLayer",
     "StaleGenerationError",
     "TargetGridConfig",
     "TemporalMethodConfidenceConfig",
+    "build_one_hour_training_samples",
+    "build_realtime_single_factor_layers",
+    "evaluate_persistence_baseline",
     "intake_legacy_cnn_zip",
+    "iter_training_sample_documents",
     "load_risk_build_configuration",
     "model_config_digest",
+    "readiness_as_dict",
+    "summarize_training_readiness",
+    "temporal_holdout_split",
+    "write_realtime_single_factor_outputs",
 ]
 
 __version__ = "0.2.0"